requirements.txt
discord.py==2.3.2
openai==1.10.0
requests==2.31.0
schedule==1.2.0
pytz==2023.3
pillow==11.0.0
pyyaml==6.0.1
python-dotenv==1.0.1
